{"schema":"decodex/commit/1","summary":"Split app-server dynamic tools","authority":"XY-1115"}
